+++ crates/engine/tree/src/persistence.rs
@@ -1,35 +1,322 @@
 #![allow(dead_code)]
 
-use crate::{
-    database::{DatabaseAction, DatabaseService, DatabaseServiceHandle},
-    static_files::{StaticFileAction, StaticFileService, StaticFileServiceHandle},
-};
 use reth_chain_state::ExecutedBlock;
 use reth_db::Database;
-use reth_primitives::{SealedBlock, B256, U256};
-use reth_provider::ProviderFactory;
+use reth_errors::ProviderResult;
+use reth_primitives::{SealedBlock, StaticFileSegment, TransactionSignedNoHash, B256, U256};
+use reth_provider::{
+    writer::StorageWriter, BlockExecutionWriter, BlockNumReader, BlockWriter, HistoryWriter,
+    OriginalValuesKnown, ProviderFactory, StageCheckpointWriter, StateWriter,
+    StaticFileProviderFactory, StaticFileWriter, TransactionsProviderExt,
+};
 use reth_prune::{Pruner, PrunerOutput};
+use reth_stages_types::{StageCheckpoint, StageId};
 use std::sync::{
-    mpsc::{SendError, Sender},
+    mpsc::{Receiver, SendError, Sender},
     Arc,
 };
 use tokio::sync::oneshot;
+use tracing::debug;
 
-/// A signal to the database and static file services that part of the tree state can be persisted.
+/// Writes parts of reth's in memory tree state to the database and static files.
+///
+/// This is meant to be a spawned service that listens for various incoming persistence operations,
+/// performing those actions on disk, and returning the result in a channel.
+///
+/// There are two types of operations this service can perform:
+/// - Writing executed blocks to disk, returning the hash of the latest block that was inserted.
+/// - Removing blocks from disk, returning the hash of the lowest block removed.
+///
+/// This should be spawned in its own thread with [`std::thread::spawn`], since this performs
+/// blocking I/O operations in an endless loop.
+#[derive(Debug)]
+pub struct PersistenceService<DB> {
+    /// The provider factory to use
+    provider: ProviderFactory<DB>,
+    /// Incoming requests
+    incoming: Receiver<PersistenceAction>,
+    /// The pruner
+    pruner: Pruner<DB, ProviderFactory<DB>>,
+}
+
+impl<DB: Database> PersistenceService<DB> {
+    /// Create a new persistence service
+    pub const fn new(
+        provider: ProviderFactory<DB>,
+        incoming: Receiver<PersistenceAction>,
+        pruner: Pruner<DB, ProviderFactory<DB>>,
+    ) -> Self {
+        Self { provider, incoming, pruner }
+    }
+
+    /// Writes the cloned tree state to database
+    fn write(&self, blocks: Vec<ExecutedBlock>) -> ProviderResult<()> {
+        let provider_rw = self.provider.provider_rw()?;
+
+        if blocks.is_empty() {
+            debug!(target: "tree::persistence::db", "Attempted to write empty block range");
+            return Ok(())
+        }
+
+        let first_number = blocks.first().unwrap().block().number;
+
+        let last = blocks.last().unwrap().block();
+        let last_block_number = last.number;
+
+        // TODO: remove all the clones and do performant / batched writes for each type of object
+        // instead of a loop over all blocks,
+        // meaning:
+        //  * blocks
+        //  * state
+        //  * hashed state
+        //  * trie updates (cannot naively extend, need helper)
+        //  * indices (already done basically)
+        // Insert the blocks
+        for block in blocks {
+            let sealed_block =
+                block.block().clone().try_with_senders_unchecked(block.senders().clone()).unwrap();
+            provider_rw.insert_block(sealed_block)?;
+
+            // Write state and changesets to the database.
+            // Must be written after blocks because of the receipt lookup.
+            let execution_outcome = block.execution_outcome().clone();
+            // TODO: do we provide a static file producer here?
+            execution_outcome.write_to_storage(&provider_rw, None, OriginalValuesKnown::No)?;
+
+            // insert hashes and intermediate merkle nodes
+            {
+                let trie_updates = block.trie_updates().clone();
+                let hashed_state = block.hashed_state();
+                // TODO: use single storage writer in task when sf / db tasks are combined
+                let storage_writer = StorageWriter::new(Some(&provider_rw), None);
+                storage_writer.write_hashed_state(&hashed_state.clone().into_sorted())?;
+                trie_updates.write_to_database(provider_rw.tx_ref())?;
+            }
+
+            // update history indices
+            provider_rw.update_history_indices(first_number..=last_block_number)?;
+
+            // Update pipeline progress
+            provider_rw.update_pipeline_stages(last_block_number, false)?;
+        }
+
+        debug!(target: "tree::persistence::db", range = ?first_number..=last_block_number, "Appended blocks");
+
+        Ok(())
+    }
+
+    /// Removes block data above the given block number from the database.
+    /// This is exclusive, i.e., it only removes blocks above `block_number`, and does not remove
+    /// `block_number`.
+    ///
+    /// This will then send a command to the static file service, to remove the actual block data.
+    fn remove_blocks_above(&self, block_number: u64) -> ProviderResult<()> {
+        let provider_rw = self.provider.provider_rw()?;
+        let highest_block = self.provider.last_block_number()?;
+        provider_rw.remove_block_and_execution_range(block_number..=highest_block)?;
+
+        Ok(())
+    }
+
+    /// Prunes block data before the given block hash according to the configured prune
+    /// configuration.
+    fn prune_before(&mut self, block_num: u64) -> PrunerOutput {
+        // TODO: doing this properly depends on pruner segment changes
+        self.pruner.run(block_num).expect("todo: handle errors")
+    }
+
+    /// Updates checkpoints related to block headers and bodies. This should be called after new
+    /// transactions have been successfully written to disk.
+    fn update_transaction_meta(&self, block_num: u64) -> ProviderResult<()> {
+        let provider_rw = self.provider.provider_rw()?;
+        provider_rw.save_stage_checkpoint(StageId::Headers, StageCheckpoint::new(block_num))?;
+        provider_rw.save_stage_checkpoint(StageId::Bodies, StageCheckpoint::new(block_num))?;
+        provider_rw.commit()?;
+        Ok(())
+    }
+
+    /// Writes the transactions to static files, to act as a log.
+    ///
+    /// The [`update_transaction_meta`](Self::update_transaction_meta) method should be called
+    /// after this, to update the checkpoints for headers and block bodies.
+    fn log_transactions(
+        &self,
+        block: Arc<SealedBlock>,
+        start_tx_number: u64,
+        td: U256,
+    ) -> ProviderResult<u64> {
+        let provider = self.provider.static_file_provider();
+        let mut header_writer = provider.get_writer(block.number, StaticFileSegment::Headers)?;
+        let mut transactions_writer =
+            provider.get_writer(block.number, StaticFileSegment::Transactions)?;
+
+        header_writer.append_header(block.header(), td, &block.hash())?;
+        let no_hash_transactions =
+            block.body.clone().into_iter().map(TransactionSignedNoHash::from);
+
+        let mut tx_number = start_tx_number;
+        for tx in no_hash_transactions {
+            transactions_writer.append_transaction(tx_number, &tx)?;
+            tx_number += 1;
+        }
+
+        // increment block for transactions
+        transactions_writer.increment_block(StaticFileSegment::Transactions, block.number)?;
+
+        // finally commit
+        transactions_writer.commit()?;
+        header_writer.commit()?;
+
+        Ok(block.number)
+    }
+
+    /// Write execution-related block data to static files.
+    ///
+    /// This will then send a command to the db service, that it should write new data, and update
+    /// the checkpoints for execution and beyond.
+    fn write_execution_data(&self, blocks: &[ExecutedBlock]) -> ProviderResult<()> {
+        if blocks.is_empty() {
+            return Ok(())
+        }
+        let provider_rw = self.provider.provider_rw()?;
+        let provider = self.provider.static_file_provider();
+
+        // NOTE: checked non-empty above
+        let first_block = blocks.first().unwrap().block();
+        let last_block = blocks.last().unwrap().block().clone();
+
+        // use the storage writer
+        let current_block = first_block.number;
+        let receipts_writer =
+            provider.get_writer(first_block.number, StaticFileSegment::Receipts)?;
+
+        let storage_writer = StorageWriter::new(Some(&provider_rw), Some(receipts_writer));
+        let receipts_iter = blocks.iter().map(|block| {
+            let receipts = block.execution_outcome().receipts().receipt_vec.clone();
+            debug_assert!(receipts.len() == 1);
+            receipts.first().unwrap().clone()
+        });
+        storage_writer.append_receipts_from_blocks(current_block, receipts_iter)?;
+
+        Ok(())
+    }
+
+    /// Removes the blocks above the given block number from static files. Also removes related
+    /// receipt and header data.
+    ///
+    /// This is exclusive, i.e., it only removes blocks above `block_number`, and does not remove
+    /// `block_number`.
+    ///
+    /// Returns the block hash for the lowest block removed from the database, which should be
+    /// the hash for `block_number + 1`.
+    ///
+    /// This is meant to be called by the db service, as this should only be done after related data
+    /// is removed from the database, and checkpoints are updated.
+    ///
+    /// Returns the hash of the lowest removed block.
+    fn remove_static_file_blocks_above(&self, block_num: u64) -> ProviderResult<()> {
+        let sf_provider = self.provider.static_file_provider();
+        let db_provider_rw = self.provider.provider_rw()?;
+
+        // get highest static file block for the total block range
+        let highest_static_file_block = sf_provider
+            .get_highest_static_file_block(StaticFileSegment::Headers)
+            .expect("todo: error handling, headers should exist");
+
+        // Get the total txs for the block range, so we have the correct number of columns for
+        // receipts and transactions
+        let tx_range = db_provider_rw
+            .transaction_range_by_block_range(block_num..=highest_static_file_block)?;
+        let total_txs = tx_range.end().saturating_sub(*tx_range.start());
+
+        // get the writers
+        let mut header_writer = sf_provider.get_writer(block_num, StaticFileSegment::Headers)?;
+        let mut transactions_writer =
+            sf_provider.get_writer(block_num, StaticFileSegment::Transactions)?;
+        let mut receipts_writer = sf_provider.get_writer(block_num, StaticFileSegment::Receipts)?;
+
+        // finally actually truncate, these internally commit
+        receipts_writer.prune_receipts(total_txs, block_num)?;
+        transactions_writer.prune_transactions(total_txs, block_num)?;
+        header_writer.prune_headers(highest_static_file_block.saturating_sub(block_num))?;
+
+        sf_provider.commit()?;
+
+        Ok(())
+    }
+}
+
+impl<DB> PersistenceService<DB>
+where
+    DB: Database,
+{
+    /// This is the main loop, that will listen to database events and perform the requested
+    /// database actions
+    pub fn run(mut self) {
+        // If the receiver errors then senders have disconnected, so the loop should then end.
+        while let Ok(action) = self.incoming.recv() {
+            match action {
+                PersistenceAction::RemoveBlocksAbove((new_tip_num, sender)) => {
+                    self.remove_blocks_above(new_tip_num).expect("todo: handle errors");
+                    self.remove_static_file_blocks_above(new_tip_num).expect("todo: handle errors");
+
+                    // we ignore the error because the caller may or may not care about the result
+                    let _ = sender.send(());
+                }
+                PersistenceAction::SaveBlocks((blocks, sender)) => {
+                    if blocks.is_empty() {
+                        todo!("return error or something");
+                    }
+                    let last_block_hash = blocks.last().unwrap().block().hash();
+                    // first write to db
+                    self.write_execution_data(&blocks).expect("todo: handle errors");
+                    // then write to static files
+                    self.write(blocks).expect("todo: handle errors");
+
+                    // we ignore the error because the caller may or may not care about the result
+                    let _ = sender.send(last_block_hash);
+                }
+                PersistenceAction::PruneBefore((block_num, sender)) => {
+                    let res = self.prune_before(block_num);
+
+                    // we ignore the error because the caller may or may not care about the result
+                    let _ = sender.send(res);
+                }
+                PersistenceAction::LogTransactions((block, start_tx_number, td, sender)) => {
+                    let block_num = self
+                        .log_transactions(block, start_tx_number, td)
+                        .expect("todo: handle errors");
+                    self.update_transaction_meta(block_num).expect("todo: handle errors");
+
+                    // we ignore the error because the caller may or may not care about the result
+                    let _ = sender.send(());
+                }
+            }
+        }
+    }
+}
+
+/// A signal to the persistence service that part of the tree state can be persisted.
 #[derive(Debug)]
 pub enum PersistenceAction {
-    /// The given block has been added to the canonical chain, its transactions and headers will be
-    /// persisted for durability.
-    LogTransactions((Arc<SealedBlock>, u64, U256, oneshot::Sender<()>)),
-
     /// The section of tree state that should be persisted. These blocks are expected in order of
     /// increasing block number.
     ///
-    /// This should just store the execution history-related data. Header, transaction, and
-    /// receipt-related data should already be written to static files.
+    /// First, header, transaction, and receipt-related data should be written to static files.
+    /// Then the execution history-related data will be written to the database.
     SaveBlocks((Vec<ExecutedBlock>, oneshot::Sender<B256>)),
 
+    /// The given block has been added to the canonical chain, its transactions and headers will be
+    /// persisted for durability.
+    ///
+    /// This will first append the header and transactions to static files, then update the
+    /// checkpoints for headers and block bodies in the database.
+    LogTransactions((Arc<SealedBlock>, u64, U256, oneshot::Sender<()>)),
+
     /// Removes block data above the given block number from the database.
+    ///
+    /// This will first update checkpoints from the database, then remove actual block data from
+    /// static files.
     RemoveBlocksAbove((u64, oneshot::Sender<()>)),
 
     /// Prune associated block data before the given block number, according to already-configured
@@ -37,66 +324,35 @@ pub enum PersistenceAction {
     PruneBefore((u64, oneshot::Sender<PrunerOutput>)),
 }
 
-/// A handle to the database and static file services. This will send commands to the correct
-/// service, depending on the command.
-///
-/// Some commands should be sent to the database service, and others should be sent to the static
-/// file service, despite having the same name. This is because some actions require work to be done
-/// by both the static file _and_ the database service, and require some coordination.
-///
-/// This type is what actually coordinates the two services, and should be used by consumers of the
-/// persistence related services.
+/// A handle to the persistence service
 #[derive(Debug, Clone)]
 pub struct PersistenceHandle {
-    /// The channel used to communicate with the database service
-    db_sender: Sender<DatabaseAction>,
-    /// The channel used to communicate with the static file service
-    static_file_sender: Sender<StaticFileAction>,
+    /// The channel used to communicate with the persistence service
+    sender: Sender<PersistenceAction>,
 }
 
 impl PersistenceHandle {
     /// Create a new [`PersistenceHandle`] from a [`Sender<PersistenceAction>`].
-    pub const fn new(
-        db_sender: Sender<DatabaseAction>,
-        static_file_sender: Sender<StaticFileAction>,
-    ) -> Self {
-        Self { db_sender, static_file_sender }
+    pub const fn new(sender: Sender<PersistenceAction>) -> Self {
+        Self { sender }
     }
 
-    /// Create a new [`PersistenceHandle`], and spawn the database and static file services.
+    /// Create a new [`PersistenceHandle`], and spawn the persistence service.
     pub fn spawn_services<DB: Database + 'static>(
         provider_factory: ProviderFactory<DB>,
         pruner: Pruner<DB, ProviderFactory<DB>>,
     ) -> Self {
         // create the initial channels
-        let (static_file_service_tx, static_file_service_rx) = std::sync::mpsc::channel();
         let (db_service_tx, db_service_rx) = std::sync::mpsc::channel();
 
         // construct persistence handle
-        let persistence_handle = Self::new(db_service_tx.clone(), static_file_service_tx.clone());
+        let persistence_handle = Self::new(db_service_tx);
 
-        // construct handles for the services to talk to each other
-        let static_file_handle = StaticFileServiceHandle::new(static_file_service_tx);
-        let database_handle = DatabaseServiceHandle::new(db_service_tx);
-
-        // spawn the db service
-        let db_service = DatabaseService::new(
-            provider_factory.clone(),
-            db_service_rx,
-            static_file_handle,
-            pruner,
-        );
+        // spawn the persistence service
+        let db_service = PersistenceService::new(provider_factory, db_service_rx, pruner);
         std::thread::Builder::new()
-            .name("Database Service".to_string())
+            .name("Persistence Service".to_string())
             .spawn(|| db_service.run())
-            .unwrap();
-
-        // spawn the static file service
-        let static_file_service =
-            StaticFileService::new(provider_factory, static_file_service_rx, database_handle);
-        std::thread::Builder::new()
-            .name("Static File Service".to_string())
-            .spawn(|| static_file_service.run())
             .unwrap();
 
         persistence_handle
@@ -104,23 +360,11 @@ impl PersistenceHandle {
 
     /// Sends a specific [`PersistenceAction`] in the contained channel. The caller is responsible
     /// for creating any channels for the given action.
-    pub fn send_action(&self, action: PersistenceAction) -> Result<(), SendError<DatabaseAction>> {
-        match action {
-            PersistenceAction::LogTransactions(input) => self
-                .db_sender
-                .send(DatabaseAction::LogTransactions(input))
-                .map_err(From::from),
-            PersistenceAction::SaveBlocks(input) => self
-                .db_sender
-                .send(DatabaseAction::SaveBlocks(input))
-                .map_err(From::from),
-            PersistenceAction::RemoveBlocksAbove(input) => {
-                self.db_sender.send(DatabaseAction::RemoveBlocksAbove(input)).map_err(From::from)
-            }
-            PersistenceAction::PruneBefore(input) => {
-                self.db_sender.send(DatabaseAction::PruneBefore(input)).map_err(From::from)
-            }
-        }
+    pub fn send_action(
+        &self,
+        action: PersistenceAction,
+    ) -> Result<(), SendError<PersistenceAction>> {
+        self.sender.send(action)
     }
 
     /// Tells the persistence service to save a certain list of finalized blocks. The blocks are
